backend/requirements.txt
Flask==2.2.2
flask-jwt-extended==4.3.1
Flask-SQLAlchemy==2.5.1
flask-cors==3.0.10
pytest==7.0.0
Werkzeug==2.2.2
python-dotenv==1.0.1